chore: move LocaleDefinition to definitions
